feat(ws_metrics): WS metrics
