+++ pragma-node/src/utils/mod.rs
@@ -22,7 +22,6 @@ mod aws;
 mod conversion;
 mod custom_extractors;
 pub mod doc_examples;
-pub mod pricing;
 mod signing;
 
 const ONE_YEAR_IN_SECONDS: f64 = 3153600_f64;
